// src/app/lib/generated/UserChange.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Role } from "./Role";

/**
 * One change to an account.
 */
export type UserChange = { "type": "rename", name: string, } | { "type": "setPassword", password: string | null, } | { "type": "setRole", role: Role, };